Extract StructLikeWrapperWithFieldIdToIndex to separate class
